[fix] .git 파일 폴더 내로 이동
