--- HamGaja/src/pages/Detail.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Detail() {
-  return <div>Detail</div>
-}
-
-export default Detail
