--- HamGaja/src/pages/Login.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Login() {
-  return <div>Login</div>
-}
-
-export default Login
