--- HamGaja/src/pages/NotFound.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function NotFound() {
-  return <div>NotFound</div>
-}
-
-export default NotFound
