--- HamGaja/src/pages/Post.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Post() {
-  return <div>Post</div>
-}
-
-export default Post
