--- HamGaja/src/pages/Register.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-function Register() {
-  return <div>Register</div>
-}
-
-export default Register
