+++ README.md
@@ -1,1 +1,3 @@
-# Frontend
+# 프로젝트 할때 땡겨가라.. 무료나눔 한다....
+
+## 도움이 되었다면 디엠으로 감사인사 정도는 해라......
