+++ src/pages/Detail.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Detail() {
+  return <div>Detail</div>
+}
+
+export default Detail
