+++ src/pages/NotFound.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function NotFound() {
+  return <div>NotFound</div>
+}
+
+export default NotFound
