+++ src/pages/Post.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Post() {
+  return <div>Post</div>
+}
+
+export default Post
