+++ src/pages/Register.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Register() {
+  return <div>Register</div>
+}
+
+export default Register
